# api/docs/swagger.yaml
definitions:
  models.CreateUserRoleRequest:
    properties:
      method:
        type: string
      path:
        type: string
      role_name:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/rbac/add-user-role:
    post:
      consumes:
      - application/json
      description: Create new user-role
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateUserRoleRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Create new user-role
      tags:
      - rbac
  /v1/rbac/policy:
    get:
      consumes:
      - application/json
      description: This API gets list of policies
      parameters:
      - description: Role
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get list of policeis
      tags:
      - rbac
  /v1/rbac/roles:
    get:
      consumes:
      - application/json
      description: This API get list of roles
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Get list of roles
      tags:
      - rbac
...
swagger: "2.0"